Object Collisions Working Great + Platforms
Adding in objects as collisions as worked great. I've also gone and dynamically deleted the tops of every tileset and replaced them with another object and very simplified the collision code.

Now adding in platforms. X and Y work fine, but Z is trickier because the player can be under it.